add layout for edit

+++ app/Http/Controllers/CategoryController.php
@@ -15,9 +15,16 @@ class CategoryController extends Controller
     public function getCategoryTemplateForAdmin($id = null, Request $request)
     {
         if ($request->user()->can('edit', Category::class)) {
-            return view('admin.categoryTemplate');
+            $category = Category::find($id);
+            if (!empty($category)) {
+                return view('admin.categoryTemplate', [
+                    'category' => $category
+                ]);
+            } else {
+                abort(404);
+            }
         } else {
-            abort(404);
+            abort(403);
         }
     }
 
@@ -25,17 +32,17 @@ class CategoryController extends Controller
     {
         $data = $request->all();
         $data['parent_id'] = strip_tags($data['parent_id']);
-        $exclude_id = [];
-        if (!empty($data['exclude_id'])) {
-            foreach ($data['exclude_id'] as $item) {
-                $exclude_id[] = strip_tags($item);
+        $exclude_ids = [];
+        if (!empty($data['exclude_ids'])) {
+            foreach ($data['exclude_ids'] as $item) {
+                $exclude_ids[] = strip_tags($item);
             }
         }
         $categories = Category::all();
         foreach ($categories as $category) {
             $category->append(array('routes'));
         }
-        $categories = build_tree($categories, $data['parent_id'], $exclude_id);
+        $categories = build_tree($categories, $data['parent_id'], $exclude_ids);
         return response()->json($categories);
     }
 
@@ -43,16 +50,20 @@ class CategoryController extends Controller
     {
         if ($request->user()->can('delete', Category::class)) {
             $category = Category::find($id);
-            $new_parent_cat_id = 0;
-            if(!empty($category->parentCategory->id)){
-                $new_parent_cat_id = $category->parentCategory->id;
+            if (!empty($category)) {
+                $new_parent_cat_id = 0;
+                if (!empty($category->parentCategory->id)) {
+                    $new_parent_cat_id = $category->parentCategory->id;
+                }
+                foreach ($category->childrenCategories as $child) {
+                    $child->parent = $new_parent_cat_id;
+                    $child->save();
+                }
+                $category->books()->detach();
+                $category->delete();
+                return response()->json(['status' => 'success']);
             }
-            foreach ($category->childrenCategories as $child){
-                $child->parent = $new_parent_cat_id;
-                $child->save();
-            }
-            $category->books()->detach();
-            $category->delete();
+            abort(404);
         } else {
             abort(403);
         }
